Added const helpers for easy evidence
All of the evidence for typelcasses has to be provided to the typeclass
functions (in Scala, they just use implicits everywhere, which we don't
have in Rust, so they have to be specified).  To make this less
frustrating, provide  a const for each of the effects evidence structs
to make passing it in easier.

Also add a `sg(ev)` func which takes inner evidence for Semigroups which
have inner types (liek options, results, futures, etc.) in order to
provide how the user wants to combine the inner data (add, multiply,
append, etc.).

These are references (except SG consts which are plain structs) to make
usage easier.