Remove the self from all typeclasses
It's no longer needed, as typeclasses don't need to be trait objects
anymore, so it's just extra noise.  This also lets us get rid of the
dispatcher on all the *Effect classes, which just returned its own "Fct"
type as an actual struct.  Now, we jsut need the static type to call teh
static functions to do the different typeclass work.