Allow speific inner semigroup to be used
When the default of adding two ints (e.g) , or if other defaults for various
types aren't good enough, use `combine_inner::<_, _, SGTtype>()`, and
give it the SgType of the semigroup you wish to use ("IntMulSemigroup"
for example).  The first two parameters can be elided, because they are
inferred from the parameters.