Implement SyncT, flesh out IO Monad
Implement SyncT (Sync is already taken by Rust core), to allow for
`delay` and `suspend`.  The former takes a code block which takes no
parameters and returns a concrete data type (known as a thunk).  The
latter takes no parameters and returns a type constructor (such as
Option, Future, IO, etc.).  The `delay` function is synonymous to
calling `suspend` with `pure` on the paramter (to generate a type
constructor for X).

The main difference between using SyncT's `delay` and using `pure`
from Applicative is that the former will cause the thunk to be evaluated
only when commanded by the context (for example, using IO's `run_sync()`
call), while the latter is evaluated greedily, meaning as soon as the
`pure` function is called.

This also forms the main difference between Future and IO, the latter
implements SyncT, while the former does not, meaning IO effects can
always be postponed until `run_sync` is called.

Updated IO monad example to show the power of IO and SyncT.