Add "Effectful" trait as a helper
Since the effect traits are all separated, specifying a type which is an
ApplicativeEffect, FunctorEffect, MonadEffect, MonadErrorEffect, and
SyncTEffect can be cumbersome.  "Effectful" is a way to tie all of these
together, specifying a type is a type constructor which has an effect
implemented for all of the above typeclasses.

This allows easy passing of generics into functions which need a type
bound of the above effects with a certain type as the inner type (the
construction-enabling type (such as "u32" in "Option<u32>")), a certain
error type for MonadError, and an optional set of mapped type/type
constructors (in the case opf fmap or flat_map, where an "output" type
is needed to describe the effect).

Also updated the io_monad example to show use of the typeclasses
directly (as evidence which musst be passed in)-- this is closer to the
cats-style, although without higher-kinded-types, it makes the
declarations fairly verbose-- and using effects with Effectful, which
allows inference of the typeclasses based on related effects (this makes
the declarations a lot smoother, although it relies on the type
inference system to do the heavy lifting).